Redemption Prayer/Fixes/Clues/Tidying up/Rares (#207)
- closes #205
- closes #197
- closes #145

Does the followiwng
- Adds functional redemption prayer
- Fixes max stack deletion when trading
- Adds toggle for clue scrolls, currently set on disabled until someone gets the time to finish clue scrolls
- Cleans up a few minor things
- Removes a few bits of dead code
- Re enables rare detection

+++ 2006Redone Server/src/redone/net/packets/impl/ClickItem.java
@@ -1,5 +1,6 @@
 package redone.net.packets.impl;
 
+import redone.Constants;
 import redone.Server;
 import redone.game.content.consumables.Beverages;
 import redone.game.content.consumables.Kebabs;
@@ -143,16 +144,19 @@ public class ClickItem implements PacketType {
 			break;
 			
 		case 2677:
+			if (Constants.CLUES_ENABLED)
 			player.getItemAssistant().deleteItem(itemId, 1);
 			TreasureTrails.addClueReward(player, 0);
 			break;
 
 		case 2678:
+			if (Constants.CLUES_ENABLED)
 			player.getItemAssistant().deleteItem(itemId, 1);
 			TreasureTrails.addClueReward(player, 1);
 			break;
 
 		case 2679:
+			if (Constants.CLUES_ENABLED)
 			player.getItemAssistant().deleteItem(itemId, 1);
 			TreasureTrails.addClueReward(player, 2);
 			break;
